+++ system_info.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 
+# Ask if the user wants to collect sensor data
+echo ""
+read -p "Do you want to collect sensor data? (yes or no, default is yes): " sensors
+
+# Setting default value
+if [ "$sensors" = "" ]; then
+	sensors="yes"
+fi
+
 # Starting sensor data collection
-sh sensor_data.sh &
-sensor_process=$!
+if [ "$sensors" = "yes" ]; then
+	sh sensor_data.sh &
+	sensor_process=$!
+fi
 
 # Script Start Date and Time (for use in file name)
 date=`date +"%m-%d-%y_%T"`
@@ -33,8 +44,6 @@ echo "kernel=`uname -r`" >> $file
 #BIOS Info
 echo "\n\n\n\nBIOS INFO:" >> $file
 sudo dmidecode --type bios >> $file
-#echo "bios version=`sudo dmidecode -s bios-version`" >> $file
-#echo "bios release date=`sudo dmidecode -s bios-release date`" >> $file
 
 # BMC Info
 echo "\n\n\n\nBMC INFO:" >> $file
@@ -68,6 +77,7 @@ perl excel_conv.pl "$file" "" "$date"
 # Deleting the temp files needed for the excel files after they are inserted
 rm sys_topo_${date}.png
 
-
 # Ending sensor data collection
-sudo kill $sensor_process
+if [ "$sensors" = "yes" ]; then
+	sudo kill $sensor_process
+fi
